requirements.txt
beautifulsoup4~=4.12.2
requests~=2.31.0
pyyaml~=6.0.1